Add files via upload
I named my app "ConvertEase." It’s a simple calculator for converting USD dollars and temperatures. The reason I created this app is pretty personal for me . My brother lives in the US, and he always talks about prices in dollars, which gets confusing for me since I’m used to thinking in rupees. I found myself Googling all the time just to convert dollars to rupees. And with temperatures too . So, I thought, why not make an app that does it for me? That’s how ConvertEase is done!